up proxies

+++ tools/update.ps1
@@ -1,5 +1,6 @@
 param(
-    [string]$Proxies = "http://45.136.131.31:8443,http://45.136.131.42:8447,http://138.124.53.25:7443,http://38.145.218.10:8443"
+    [string]$Proxies = "http://121.126.185.63:25152,http://38.145.203.135:8443,http://216.180.127.45:1080,http://85.198.96.242:3128,http://38.145.218.82:8443,http://45.136.130.216:8443,http://103.30.30.226:20326,http://45.136.130.211:8447",
+    [switch]$Clean = $false
 )
 
 #region Dependencies
@@ -15,13 +16,15 @@ Write-Host "Logging to $LogFile" -ForegroundColor DarkGray
 Start-Transcript -Path $LogFile -Append -Force | Out-Null
 
 try {
-    Write-Host "Cleaning profiles: $ProfilesDir" -ForegroundColor Yellow
-    if (Test-Path $ProfilesDir) {
-        Get-ChildItem -Path $ProfilesDir -Directory | Remove-Item -Recurse -Force -ErrorAction SilentlyContinue 2>$null
+    if ($Clean) {
+        Write-Host "Cleaning profiles: $ProfilesDir" -ForegroundColor Yellow
+        if (Test-Path $ProfilesDir) {
+            Get-ChildItem -Path $ProfilesDir -Directory | Remove-Item -Recurse -Force -ErrorAction SilentlyContinue 2>$null
+        }
     }
-    .\tools\Update-FromDiscord.ps1 -Extract -Debug -Proxies $Proxies
-    .\tools\Update-FromUEVRDeluxe.ps1 -Fetch -Download -Extract -CleanCache -CleanDownloads -Debug -Proxies $Proxies
-    .\tools\Update-FromUEVRProfiles.ps1 -Fetch -Download -Extract -CleanCache -CleanDownloads -Debug -Proxies $Proxies
+    .\tools\Update-FromDiscord.ps1 -Extract -Debug
+    .\tools\Update-FromUEVRDeluxe.ps1 -Extract -CleanCache -CleanDownloads -Debug -Proxies $Proxies
+    .\tools\Update-FromUEVRProfiles.ps1 -Download -Extract -CleanCache -CleanDownloads -Debug -Proxies $Proxies
     
     .\tools\Find-Issues.ps1 -Fix -Debug
     .\tools\Process-Whitelist.ps1 -Delete -Debug
